# mindmate-ai/technical/api_spec.yaml
openapi: 3.0.3
info:
  title: MindMate AI API
  description: |
    # MindMate AI REST API
    
    Production-ready API specification for MindMate AI - An AI-powered mental health support platform.
    
    ## Features
    - **Authentication**: JWT-based authentication with refresh tokens
    - **AI Therapy Sessions**: Real-time conversational AI therapy
    - **Mood Tracking**: Daily mood logging and pattern analysis
    - **Task Management**: Personalized wellness tasks and goals
    - **Crisis Support**: Emergency resources and immediate assistance
    - **Progress Analytics**: Insights and streak tracking
    - **Notifications**: Smart reminders and wellness nudges
    
    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```
    
    ## Rate Limits
    - **Anonymous**: 10 requests/minute
    - **Authenticated**: 100 requests/minute
    - **Premium**: 500 requests/minute
    - **Admin**: 1000 requests/minute
    
    ## Error Codes
    | Code | Description |
    |------|-------------|
    | 400 | Bad Request - Invalid input |
    | 401 | Unauthorized - Invalid or missing token |
    | 403 | Forbidden - Insufficient permissions |
    | 404 | Not Found - Resource doesn't exist |
    | 409 | Conflict - Resource already exists |
    | 422 | Unprocessable Entity - Validation error |
    | 429 | Too Many Requests - Rate limit exceeded |
    | 500 | Internal Server Error |
  
...
servers:
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account with email/password
      operationId: registerUser
      security: []
      x-rate-limit: 5/minute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT tokens
      operationId: loginUser
      security: []
      x-rate-limit: 10/minute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Account suspended or not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/refresh endpoints
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: bad_request
            message: Invalid request parameters
            code: 400
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Invalid or missing authentication token
            code: 401
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: conflict
            message: Resource already exists
            code: 409
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
        - code
      properties:
        error:
          type: string
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        code:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error details
        request_id:
          type: string
          description: Unique request ID for debugging
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            field_errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string

    # ==================== AUTH SCHEMAS ====================
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          description: Must contain at least one uppercase, one lowercase, one number
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        date_of_birth:
          type: string
          format: date
        timezone:
          type: string
          default: UTC
        referral_code:
          type: string
          description: Optional referral code
        marketing_consent:
          type: boolean
          default: false
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        device_info:
          type: object
          properties:
            device_id:
              type: string
            platform:
              type: string
              enum: [ios, android, web]
            app_version:
              type: string
    AuthResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - user
      properties:
        access_token:
          type: string
          description: JWT access token (expires in 15 minutes)
        refresh_token:
          type: string
          description: Refresh token (expires in 30 days)
        expires_in:
          type: integer
          description: Access token expiration time in seconds
          default: 900
        token_type:
          type: string
          default: Bearer
        user:
          $ref: '#/components/schemas/UserProfile'
        mfa_required:
          type: boolean
          default: false
        temp_token:
          type: string
          description: Temporary token when MFA is required
...
    UserProfile:
      type: object
      required:
        - id
        - email
        - created_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        avatar_url:
          type: string
          format: uri
        date_of_birth:
          type: string
          format: date
        timezone:
          type: string
        email_verified:
          type: boolean
        mfa_enabled:
          type: boolean
        subscription_tier:
          type: string
          enum: [free, premium, enterprise]
        subscription_status:
          type: string
          enum: [active, cancelled, expired, pending]
        onboarding_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
        account_status:
          type: string
          enum: [active, suspended, pending_deletion]
...